[F] Integrate saml providers into identity record
